FutureUtil / StockOptionUtil / StockOptionService: case Duration to double
INSERT_INTO_TICK_SIMULATED_OPTIONS_BY_IVOL: set scale

git-svn-id: https://repo.algotrader.ch/svn/trunk/AlgoTrader@5707 19eef36c-86f0-de11-afab-0004ac968479

+++ common/src/main/java/com/algoTrader/future/FutureUtil.java
@@ -33,7 +33,7 @@ public class FutureUtil {
 
         FutureFamily family = (FutureFamily) future.getSecurityFamily();
 
-        double years = (future.getExpiration().getTime() - DateUtil.getCurrentEPTime().getTime()) / Duration.YEAR_1.getValue();
+        double years = (future.getExpiration().getTime() - DateUtil.getCurrentEPTime().getTime()) / (double) Duration.YEAR_1.getValue();
 
         return getFuturePrice(underlyingSpot, years, family.getIntrest(), family.getDividend());
     }
