+++ common/src/main/java/com/algoTrader/stockOption/StockOptionUtil.java
@@ -68,7 +68,7 @@ public class StockOptionUtil {
 
         StockOptionFamily family = (StockOptionFamily) stockOption.getSecurityFamily();
 
-        double years = (stockOption.getExpiration().getTime() - DateUtil.getCurrentEPTime().getTime()) / Duration.YEAR_1.getValue();
+        double years = (stockOption.getExpiration().getTime() - DateUtil.getCurrentEPTime().getTime()) / (double) Duration.YEAR_1.getValue();
 
         return getOptionPrice(underlyingSpot, stockOption.getStrike().doubleValue(), vola, years, family.getIntrest(), family.getDividend(), stockOption.getType());
     }
@@ -104,7 +104,7 @@ public class StockOptionUtil {
 
         StockOptionFamily family = (StockOptionFamily) stockOption.getSecurityFamily();
 
-        double years = (stockOption.getExpiration().getTime() - DateUtil.getCurrentEPTime().getTime()) / Duration.YEAR_1.getValue();
+        double years = (stockOption.getExpiration().getTime() - DateUtil.getCurrentEPTime().getTime()) / (double) Duration.YEAR_1.getValue();
 
         return getImpliedVolatility(underlyingSpot, stockOption.getStrike().doubleValue(), currentValue, years, family.getIntrest(), family.getDividend(),
                 stockOption.getType());
@@ -142,7 +142,7 @@ public class StockOptionUtil {
 
         StockOptionFamily family = (StockOptionFamily) stockOption.getSecurityFamily();
 
-        double years = (stockOption.getExpiration().getTime() - DateUtil.getCurrentEPTime().getTime()) / Duration.YEAR_1.getValue();
+        double years = (stockOption.getExpiration().getTime() - DateUtil.getCurrentEPTime().getTime()) / (double) Duration.YEAR_1.getValue();
 
         return getImpliedVolatilityNR(underlyingSpot, stockOption.getStrike().doubleValue(), currentValue, years, family.getIntrest(), family.getDividend(),
                 stockOption.getType());
@@ -177,7 +177,7 @@ public class StockOptionUtil {
 
         StockOptionFamily family = (StockOptionFamily) stockOption.getSecurityFamily();
 
-        double years = (stockOption.getExpiration().getTime() - DateUtil.getCurrentEPTime().getTime()) / Duration.YEAR_1.getValue();
+        double years = (stockOption.getExpiration().getTime() - DateUtil.getCurrentEPTime().getTime()) / (double) Duration.YEAR_1.getValue();
 
         return getImpliedVolatilitySABR(underlyingSpot, stockOption.getStrike().doubleValue(), years, family.getIntrest(), family.getDividend(), stockOption.getType(), surface);
     }
@@ -217,7 +217,7 @@ public class StockOptionUtil {
         StockOptionFamily family = (StockOptionFamily) stockOption.getSecurityFamily();
 
         double strike = stockOption.getStrike().doubleValue();
-        double years = (stockOption.getExpiration().getTime() - DateUtil.getCurrentEPTime().getTime()) / Duration.YEAR_1.getValue();
+        double years = (stockOption.getExpiration().getTime() - DateUtil.getCurrentEPTime().getTime()) / (double) Duration.YEAR_1.getValue();
         double volatility = getImpliedVolatility(underlyingSpot, strike, currentValue, years, family.getIntrest(), family.getDividend(), stockOption.getType());
         return StockOptionUtil.getDelta(underlyingSpot, strike, volatility, years, family.getIntrest(), family.getDividend(), stockOption.getType());
 
@@ -238,7 +238,7 @@ public class StockOptionUtil {
         StockOptionFamily family = (StockOptionFamily) stockOption.getSecurityFamily();
 
         double strike = stockOption.getStrike().doubleValue();
-        double years = (stockOption.getExpiration().getTime() - DateUtil.getCurrentEPTime().getTime()) / Duration.YEAR_1.getValue();
+        double years = (stockOption.getExpiration().getTime() - DateUtil.getCurrentEPTime().getTime()) / (double) Duration.YEAR_1.getValue();
         double volatility = getImpliedVolatility(underlyingSpot, strike, currentValue, years, family.getIntrest(), family.getDividend(), stockOption.getType());
         return StockOptionUtil.getVega(underlyingSpot, strike, volatility, years, family.getIntrest(), family.getDividend());
     }
@@ -269,7 +269,7 @@ public class StockOptionUtil {
         StockOptionFamily family = (StockOptionFamily) stockOption.getSecurityFamily();
 
         double strike = stockOption.getStrike().doubleValue();
-        double years = (stockOption.getExpiration().getTime() - DateUtil.getCurrentEPTime().getTime()) / Duration.YEAR_1.getValue();
+        double years = (stockOption.getExpiration().getTime() - DateUtil.getCurrentEPTime().getTime()) / (double) Duration.YEAR_1.getValue();
         double volatility = getImpliedVolatility(underlyingSpot, strike, currentValue, years, family.getIntrest(), family.getDividend(), stockOption.getType());
         return StockOptionUtil.getTheta(underlyingSpot, strike, volatility, years, family.getIntrest(), family.getDividend(), stockOption.getType());
 
@@ -294,7 +294,7 @@ public class StockOptionUtil {
 
         StockOptionFamily family = (StockOptionFamily) stockOption.getSecurityFamily();
 
-        double years = (stockOption.getExpiration().getTime() - DateUtil.getCurrentEPTime().getTime()) / Duration.YEAR_1.getValue();
+        double years = (stockOption.getExpiration().getTime() - DateUtil.getCurrentEPTime().getTime()) / (double) Duration.YEAR_1.getValue();
 
         return getTotalMargin(underlyingSettlement, stockOption.getStrike().doubleValue(), stockOptionSettlement, years, family.getIntrest(),
                 family.getDividend(), stockOption.getType(), family.getMarginParameter());
